Release 1.0.0
This is the initial public release of the plugin. I have worked on this plugin for around seven months to bring this plugin to the Godot community. It is my way of saying thanks to this amazing game engine! Every piece that you need to run this is in the addon folder. (Read the README for complete instructions and documentation)